requirements.txt
-f https://pytorch-geometric.com/whl/torch-1.12.0+cu116.html
--extra-index-url https://download.pytorch.org/whl/cu116
POT
psutil
kaleido
line_profiler
black
typer
numpy
tqdm
pandas
networkx
python-igraph
matplotlib
icecream
graphviz
pygraphviz
smart-open
pydot
scipy
pytorch-lightning
torchmetrics
papermill
jupyterlab
codetiming
Pebble
NEMtropy
numba
powerlaw
optuna
giotto-tda
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
torch-geometric
torchvision
dive-into-graphs
